Add logic to book a court and to cancel booking
Add scheduleTime() action which schedules/cancels then updates
user and schedule. Export it from actions/index.js to be used
troughout the app.

Update <ScheduleModal /> with to handle scheduling.

Fix <ScheduleTime /> to be disabled when taken.

Render a message + cancel options when a user has a scheduled
court on selected day in <ScheduleBody />